requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6

SQLAlchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.3

pydantic==2.9.2
pydantic-settings==2.5.2

python-jose==3.3.0

python-multipart==0.0.12

fastapi-security==0.5.0

tzdata==2025.2